+++ src/ui/modals/mod.rs
@@ -12,8 +12,6 @@ use crate::{
 
 pub mod add_to_playlist;
 pub mod confirm_modal;
-pub mod confirm_playlist_delete;
-pub mod confirm_queue_clear;
 pub mod decoders;
 pub mod keybinds;
 pub mod outputs;